pip-requirements.txt
mesh-tensorflow